BLAZINGLY FAST!!! Added cache for GetPointerTo and a new stack allocator for the interpter

+++ src/Linux.cpp
@@ -112,7 +112,7 @@ void PlatformOutputString(string String, log_level Level)
 	write(out_fd, CLEAR, sizeof(CLEAR)-1);
 }
 
-void PlatformFreeMemory(void *Mem, uint Size)
+void PlatformFreeMemory(void *Mem, size_t Size)
 {
 	munmap(Mem, Size);
 }
