+++ src/Platform.h
@@ -17,7 +17,7 @@ typedef unsigned long (*t_proc)(void *);
 
 void PlatformWriteFile(const char *Path, u8 *Data, u32 Size); 
 string ReadEntireFile(string Path);
-void PlatformFreeMemory(void *Memory, uint Size);
+void PlatformFreeMemory(void *Memory, size_t Size);
 void *PlatformReserveMemory(size_t Size);
 void PlatformAllocateReserved(void *Memory, size_t Size);
 void PlatformOutputString(string String, log_level Level);
